Add @Producer functionality to the event bus.
A producer provides an initial-value for an event type upon subscriber regsitration. When
an subscriber method is registered on the event bus it will be called immediately with the
value from a producer if one for the associated type exists. This allows you to provide
initial values to your registering classes and guarantee that they have been initialized
after the call to register completes.