Impose restrictions on how the class hierarchy of events is treated.
For subscribers and producers, only registering with the event bus for concrete classes is allowed. You
may not register for interface types. Registration for events also now only occurs on the provided class
type itself and none of its base classes or interfaces.

When firing events, the class hierarchy is still traversed and fired to anything registered. Firing
on interfaces is no longer supported since registration for these events is also no longer supported.